Fix the problem with toggling and edit top text
I edited the top text to "switch between pen and eraser", so the message
is more clear.
There was a problem with toggling, in which by clicking on a color,
the pen would not change and you had to go from pen to eraser to color.
The problem is fixed.